requirements.txt
paho-mqtt==1.6.1
pyyaml==6.0.1
numpy==1.24.3
pytest==7.4.3
black==23.11.0
python-dateutil==2.8.2